--- a/flake.nix
+++ b/flake.nix
@@ -30,6 +30,7 @@
               # Development
               cargo-watch
               cargo-audit
+              sccache
             ] ++
 
             pkgs.lib.optionals pkgs.stdenv.isDarwin