Both on a warm cache:

Before:
cargo test test_pk_recovery  635.13s user 26.51s system 679% cpu 1:37.38 total

After:
cargo test test_pk_recovery  45.39s user 11.66s system 280% cpu 20.361 total

I should probably find out why our compile cache keeps getting invalidated, but this is a quick win. Cost is this adds in a new native dependency.